requirements.txt
# python version 3.11
# --extra-index-url https://download.pytorch.org/whl/cu124
# ninja
# torch
# torchaudio
# torchvision
# sentence_transformers
# xformers
# data
numpy
scipy
pandas
python-dotenv
contractions
nltk
# notebooks
ipython
ipykernel
ipywidgets
tqdm
# services
qdrant_client
tiktoken
openai
cohere
# backend
fastapi
uvicorn
sqlmodel
alembic
# ui
streamlit
streamlit-pdf-viewer
pymupdf
pymupdf
mammoth
Spire.PDF
Aspose.PDF
pdf2docx
tika
python-multipart
requests
fitz